ep22 done

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -33,15 +33,12 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try{
-            $this->isAble('store', Ticket::class);
+        if($this->isAble('store', Ticket::class)){
 
             return new TicketResource(Ticket::create($request->mappedAttributes()));
         }
+        return $this->error('You are not authorized for this content',403);
 
-        catch(AuthorizationException $ex){
-            return $this->error('You are not authorized for this content',403);
-        }
     }
 
     /**
@@ -71,18 +68,15 @@ class TicketController extends ApiController
         try{
             $ticket = Ticket::findOrFail($ticket_id);
 
-            //policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+                return new TicketResource($ticket);
+            }
+            return $this->error('You are not authorized for this content',403);
         }
-
         catch(ModelNotFoundException $exception){
             return $this->error('Ticket not found',404);
-        } catch(AuthorizationException $ex){
-            return $this->error('You are not authorized for this content',403);
         }
     }
 
@@ -91,11 +85,12 @@ class TicketController extends ApiController
         try{
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)){
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+                return new TicketResource($ticket);
+            }
+            return $this->error('You are not authorized for this content',403);
         }
 
         catch(ModelNotFoundException $exception){
@@ -111,11 +106,13 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('delete', $ticket);
+            if($this->isAble('delete', $ticket)) {
+                $ticket->delete();
 
-            $ticket->delete();
+                return $this->ok('Ticket successfully deleted');
+            }
 
-            return $this->ok('Ticket successfully deleted');
+            return $this->error('You are not authorized for this content',403);
 
         } catch(ModelNotFoundException $exception) {
             return $this->error('Ticket not found',404);
